+++ dev/changelog.md
@@ -2,6 +2,7 @@
 
 ## ScottPlot 4.1.21
 * Legend: Throw an exception if `RenderLegend()` is called on a plot with no labeled plottables (#1257)
+* Radar: Improved support for category labels. (#1261, #1262) _Thanks @Rayffer_
 
 ## ScottPlot 4.1.20
 * Ticks: Fixed bug where corner labels would not render when multiplier or offset notation is in use (#1252, #1253) _Thanks @@DavidBergstromSWE_